--- frontend/src/pages/brosjyre.tsx
@@ -1,13 +1,0 @@
-const BrochurePage = () => {
-    return (
-        <iframe
-            style={{
-                width: '100%',
-                height: '800px',
-            }}
-            src="https://www.visbrosjyre.no/echo/WebView/"
-        ></iframe>
-    );
-};
-
-export default BrochurePage;
